Removed nw-sendarp
The MSDN documentation for SendARP starts with: "The SendARP function
sends an Address Resolution Protocol (ARP) request to obtain the physical
address that corresponds to the specified destination IPv4 address."
However, the "Remarks" section contains the following: "If the information
requested is not in the ARP table on the local computer, then the SendARP
function will cause an ARP request to be sent to obtain the physical
address." This means that an ARP request is not sent if the IP address is
in the ARP cache of the computer. Whoops.